requirements.txt
openai-whisper
pypdf
langchain
langchain-openai
openai